src/requirements.txt
blessings==1.7
progressbar==2.5
asttokens==1.1.11
python-magic==0.4.15
rarfile==3.0
esprima==4.0.0
protobuf==3.6.1
timeout-decorator==0.4.0
func-timeout==4.3.3
networkx==2.2
enum34==1.1.6
lxml==4.2.3
pandas==0.23.4
beautifulsoup4==4.6.0
requests==2.21.0
pyyaml==3.11
